dom/indexedDB/ipc/test_ipc.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for OOP IndexedDB</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
  <body>

  <script type="application/javascript;version=1.7">
    "use strict";

    SimpleTest.waitForExplicitFinish();

    // This isn't a single test, really... It runs the entirety of the IndexedDB
    // tests. Each of those has a normal timeout handler, so there's no point in
    // having a timeout here. I'm setting this really high just to avoid getting
    // killed.
    SimpleTest.requestLongerTimeout(100);

    // Disable crash observers as it breaks later tests.
    function iframeScriptFirst() {
      SpecialPowers.prototype.registerProcessCrashObservers = function() { };
      SpecialPowers.prototype.unregisterProcessCrashObservers = function() { };

      content.wrappedJSObject.RunSet.reloadAndRunAll({
        preventDefault: function() { },
        __exposedProps__: { preventDefault: 'r' }
      });
    }

    function iframeScriptSecond() {
      let isMainProcess = content.wrappedJSObject.SpecialPowers.isMainProcess();

      sendAsyncMessage("test:indexedDB:ipcProcessType",
                       { isMainProcess: isMainProcess });

      let TestRunner = content.wrappedJSObject.TestRunner;

      let oldComplete = TestRunner.onComplete;

      TestRunner.onComplete = function() {
        TestRunner.onComplete = oldComplete;

        sendAsyncMessage("test:indexedDB:ipcTestComplete");

        if (oldComplete) {
          oldComplete();
        }
      };

      function sendTestMessage(msg) {
        sendAsyncMessage("test:indexedDB:ipcTestMessage", { msg: msg });
      }

      TestRunner.log = sendTestMessage;
      TestRunner.error = sendTestMessage;
    }

    let regexString =
      "^(TEST-PASS|TEST-UNEXPECTED-PASS|TEST-KNOWN-FAIL|TEST-UNEXPECTED-FAIL" +
      "|TEST-DEBUG-INFO|TEST-INFO) \\| ([^\\|]+) \\|(.*)";

    let regex = new RegExp(regexString);

    let seenTestMessage = false;

    function onTestMessage(data) {
      seenTestMessage = true;
      let message = SpecialPowers.wrap(data).data.msg;
      let match = regex.exec(message);
      if (match) {
        let state = match[1];
        let details = match[2] + " | " + match[3];

        switch (state) {
          case "TEST-PASS":
          case "TEST-KNOWN-FAIL":
            ok(true, details);
            break;

          case "TEST-UNEXPECTED-FAIL":
          case "TEST-UNEXPECTED-PASS":
            ok(false, details);
            break;

          case "TEST-INFO":
          case "TEST-DEBUG-INFO":
          default:
            info(details);
        }
      }
    }

    let usingChildProcess = false;

    function onProcessType(data) {
      let isMainProcess = SpecialPowers.wrap(data).data.isMainProcess;
      usingChildProcess = !isMainProcess;
    }

    function onTestComplete() {
      is(usingChildProcess, true, "Expecting to run in child process");
      is(seenTestMessage, true, "Expecting to receive messages from child");
      SpecialPowers.removePermission("browser", window.location.href);
      SimpleTest.executeSoon(function () { SimpleTest.finish(); });
    }

    function runTests() {
      let iframe = document.createElement("iframe");
      iframe.mozbrowser = true;
      iframe.id = "iframe";
      iframe.style.width = "100%";
      iframe.style.height = "1000px";

      function iframeLoadSecond() {
        ok(true, "Got second iframe load event.");
        iframe.removeEventListener("mozbrowserloadend", iframeLoadSecond);
        let mm = SpecialPowers.getBrowserFrameMessageManager(iframe);
        mm.loadFrameScript("data:,(" + iframeScriptSecond.toString() + ")();",
                           false);
      }

      function iframeLoadFirst() {
        ok(true, "Got first iframe load event.");
        iframe.removeEventListener("mozbrowserloadend", iframeLoadFirst);
        iframe.addEventListener("mozbrowserloadend", iframeLoadSecond);

        let mm = SpecialPowers.getBrowserFrameMessageManager(iframe);

        let comp = SpecialPowers.wrap(SpecialPowers.Components);

        let spObserver =
          comp.classes["@mozilla.org/special-powers-observer;1"]
              .getService(comp.interfaces.nsIMessageListener);

        mm.addMessageListener("SPPrefService", spObserver);
        mm.addMessageListener("SPProcessCrashService", spObserver);
        mm.addMessageListener("SPPingService", spObserver);
        mm.addMessageListener("SpecialPowers.Quit", spObserver);
        mm.addMessageListener("SPPermissionManager", spObserver);

        mm.addMessageListener("test:indexedDB:ipcTestMessage", onTestMessage);
        mm.addMessageListener("test:indexedDB:ipcProcessType", onProcessType);
        mm.addMessageListener("test:indexedDB:ipcTestComplete", onTestComplete);

        let specialPowersBase = "chrome://specialpowers/content/";
        mm.loadFrameScript(specialPowersBase + "MozillaLogger.js", false);
        mm.loadFrameScript(specialPowersBase + "specialpowersAPI.js", false);
        mm.loadFrameScript(specialPowersBase + "specialpowers.js", false);

        mm.loadFrameScript("data:,(" + iframeScriptFirst.toString() + ")();",
                           false);
      }

      iframe.addEventListener("mozbrowserloadend", iframeLoadFirst);

      // Strip this filename and one directory level and then add "/test".
      let href =  window.location.href;
      href = href.substring(0, href.lastIndexOf('/'));
      href = href.substring(0, href.lastIndexOf('/'));
      iframe.src = href + "/test?consoleLevel=INFO";

      document.body.appendChild(iframe);
    }

    addEventListener("load", function() {
      SpecialPowers.addPermission("browser", true, document);
      SpecialPowers.pushPrefEnv({
        "set": [
          // TODO: remove this as part of bug 820712
          ["network.disable.ipc.security", true],

          ["dom.ipc.browser_frames.oop_by_default", true],
          ["dom.mozBrowserFramesEnabled", true]
        ]
      }, runTests);
    });
  </script>
</body>
</html>
